tolto PT da main
